man up 10.10
comparing heights samples over all plots and species, calcualting stocks with tapes, letting RG and DW run through codes too

+++ scripts/03_01_LT_heights_HBI_BZE3.R
@@ -26,41 +26,37 @@ HBI_trees_removed <- read.delim(file = here(out.path, paste0((HBI_trees$inv)[1],
 # BZE3 BE dataset: 
 # this dataset contains the inventory data of the tree inventory accompanying the third national soil inventory
 # here we import a dataset called "BZE3_LT_update_2.csv" which contains plot area and stand data additionally to the original tree data
-BZE3_trees <- read.delim(file = here(out.path, "BZE3_LT_update_2.csv"), sep = ",", dec = ".")
-BZE3_trees_removed <- read.delim(file = here(out.path, paste0((BZE3_trees$inv)[1], "_LT_removed.csv")), sep = ",", dec = ".")
+try({
+  BZE3_trees <- read.delim(file = here(out.path, "BZE3_LT_update_2.csv"), sep = ",", dec = ".")
+  BZE3_trees_removed <- read.delim(file = here(out.path, paste0((BZE3_trees$inv)[1], "_LT_removed.csv")), sep = ",", dec = ".")},
+  silent = T)
 
 
-# ----- 0.6 harmonising column names & structure  -----------------------------------------------------------------
-## HBI
- HBI_trees[,c("plot_A_ha", "area_m2", "X_tree",  "Y_tree",
-             "DBH_cm", "dist_m","CCS_r_m")] <- lapply(HBI_trees[,c("plot_A_ha", "area_m2", "X_tree",  "Y_tree",
-                                                "DBH_cm", "dist_m",  "CCS_r_m")], as.numeric)
+# ----- 0.6 dataprep -----------------------------------------------------------------------
+# ----- 0.6.1. harmonising column names & structure ----------------------------------------
+if(exists('BZE3_trees') == T){
+  trees_total <-  rbind(HBI_trees, BZE3_trees %>% select(c(colnames(HBI_trees))))
+}else{
+    trees_total <- HBI_trees}
+
+## HBI & BZE3 together 
+trees_total[,c("plot_A_ha", "area_m2", "X_tree",  "Y_tree",
+             "DBH_cm", "dist_m","CCS_r_m")] <- lapply(trees_total[,c("plot_A_ha", "area_m2", "X_tree",  "Y_tree",
+                                                                   "DBH_cm", "dist_m",  "CCS_r_m")], as.numeric)
 # change -2 in H_dm and C_h_dm to NA: https://stackoverflow.com/questions/14737773/replacing-occurrences-of-a-number-in-multiple-columns-of-data-frame-with-another
-HBI_trees[,c("H_dm","C_h_dm")][HBI_trees[,c("H_dm","C_h_dm")]== -2] <- NA
-HBI_trees[,c("H_dm","C_h_dm")][HBI_trees[,c("H_dm","C_h_dm")]== -9] <- NA
+trees_total[,c("H_dm","C_h_dm")][trees_total[,c("H_dm","C_h_dm")]== -2] <- NA
+trees_total[,c("H_dm","C_h_dm")][trees_total[,c("H_dm","C_h_dm")]== -9] <- NA
 
 
-
-# BZE3_trees
-BZE3_trees[,c("plot_A_ha", "area_m2", "X_tree",  "Y_tree",
-"DBH_cm", "dist_m","CCS_r_m")] <- lapply(BZE3_trees[,c("plot_A_ha", "area_m2", "X_tree",  "Y_tree", 
-                                                                     "DBH_cm", "dist_m",  "CCS_r_m")], as.numeric)
- # change -2 in H_dm and C_h_dm to NA: https://stackoverflow.com/questions/14737773/replacing-occurrences-of-a-number-in-multiple-columns-of-data-frame-with-another
- BZE3_trees[,c("H_dm","C_h_dm")][BZE3_trees[,c("H_dm","C_h_dm")]== -2] <- NA
- BZE3_trees[,c("H_dm","C_h_dm")][BZE3_trees[,c("H_dm","C_h_dm")]== -9] <- NA
-
-
-# 1. joining in external info  -------------------------------------------------
-trees_total <- HBI_trees %>% 
-   # rbind(HBI_trees, BZE3_trees %>% 
-   #          select(c(colnames(HBI_trees)))) %>%
-# calcualte diameter and change units -----------------------------------
+# ----- 0.6.2. calcualte diameter and change units  -------------------------------------------------
+trees_total <- trees_total %>% 
   mutate(H_m = H_dm/10, 
          DBH_h_m = DBH_h_cm/100) %>%                               # change unit of DBH measuring height from cm into m by dividing by 100  
   #  apply regression of BWI (5.5.1.2.) for DBH estimation when mesasuring height differs from 1.3 m 
   mutate(DBH_class = DBH_c_function(DBH_cm, DBH_c = "class_5"), 
          BA_m2 = c_A(DBH_cm/2)*0.0001) %>% # *0.0001 to convert cm2 in m2 %>% 
  arrange(plot_ID, inv)
+
 
 # test start
 # trees_total <- rbind(trees_total %>% mutate(H_m = H_m+1.5, 
@@ -71,10 +67,8 @@ trees_total <- HBI_trees %>%
 # test end
  
  
-# 2. estimating tree height -----------------------------------------------
-
-# 2.1. fitting own models H_m ~ nls(DBH_cm) -------------------------------
-
+# 2. estimating tree height -------------------------------------------------------------
+# 2.1. fitting own models H_m ~ nls(DBH_cm) ---------------------------------------------
 # 2.1.1. non linear height model h ~ DBH per species and plot ----------------------------
 # to calculate individual tree heights for trees of the samme species and plot 
 # where the height has not been sampled we create a non-linear regression for the heights
@@ -129,7 +123,6 @@ coeff_H_SP_P <- left_join(
 # coefficents of non-linear height model per species but over all plots: 
 # https://rdrr.io/cran/forestmangr/f/vignettes/eq_group_fit_en.Rmd
 #  building separate dataframe for speicies soecific models adding adding bias, rmse and rsqrd 
-
 coeff_H_SP <- left_join(trees_total %>% 
                           select(SP_code, H_m, DBH_cm, DBH_class) %>% 
                           filter(!is.na(H_m) & !is.na(DBH_cm) & !is.na(DBH_class)) %>% 
@@ -169,13 +162,13 @@ coeff_H_SP <- left_join(trees_total %>%
   mutate(plot_ID = as.factor('all')) %>% 
   select(plot_ID, SP_code, b0, b1, b2, bias, rsme, R2, mean_h, SSres, SStot, pseu_R2, diff_h)
 
- 
-
- # 2.1.3 combined coefficients of height models ---------------------
+ # 2.1.3 combined coefficients of height models -----------------------------------------------------
 coeff_H_comb <- rbind(coeff_H_SP_P %>% mutate(plot_ID = as.factor(plot_ID)), coeff_H_SP)
 
-# 2.2 calculating Hg, Dg etc. per plot, stand, species--------------------------
- Hg_Dg_trees_total.df <- 
+
+# 2.2 calculating Hg, Dg etc. per plot, stand, species-----------------------------------------------
+# calcualte the height and diameter of a stem reprensenting the mean basal area 
+Hg_Dg_trees_total.df <- 
  trees_total %>% 
    group_by(inv, plot_ID, stand, C_layer, SP_code) %>% 
    summarise(mean_DBH_mm = mean(DBH_cm)*10,
@@ -185,14 +178,13 @@ coeff_H_comb <- rbind(coeff_H_SP_P %>% mutate(plot_ID = as.factor(plot_ID)), coe
  
   
 
-#calcualte the height and diameter of a stem reprensenting the mean basal area 
 
 
 # 2.3. height calculation -------------------------------------------------
 # 2.3.1. height calculation HBI -------------------------------------------------
 HBI_trees_update_3 <-     # this should actually be the BZE3 Datset 
   trees_total %>% 
-  filter(inv=="HBI")%>% 
+  filter(inv %in% c("HBI", "momok"))%>% ##change: we have to adjust this to select momok trees as well
 ## joining coefficients and Hg-Dg-data in
 tidyr::unite(SP_P_ID, plot_ID, SP_code, sep = "", remove = FALSE) %>%            # create column matching vectorised coefficients of coeff_SP_P (1.3. functions, h_nls_SP_P, dplyr::pull)
   left_join(.,coeff_H_SP_P %>%                                              # joining R2 from coeff_SP_P -> R2.x
@@ -223,19 +215,25 @@ tidyr::unite(SP_P_ID, plot_ID, SP_code, sep = "", remove = FALSE) %>%           
                          # and hm is na and the Slobody function cannot eb applied because there is no h_g calculatable use the curtis function
                          is.na(H_m) & is.na(R2_comb) & is.na(H_g)| is.na(H_m) & R2_comb < 0.70 & is.na(H_g) ~ h_curtis(H_SP_group, DBH_cm*10), 
                          TRUE ~ H_m))) %>% 
-   # as there were some trees that had an estimated height which was lower then the DBH measuring height. this is not only implausible but also won´t work for TapeS 
-   # thus we correct these heights afterwards by estimating their height from the relation between the dg and hg and dg and the trees DBH (dreisatz, h_proportional function)
-   mutate(H_m = ifelse(DBH_h_m > H_m, h_proportional(D_g, H_g, DBH_cm), H_m)) %>% 
-   # select columns that should enter the next step of data processing
-     select(plot_ID, inv, inv_year, stand, tree_ID,  tree_inventory_status,  multi_stem, dist_cm,  azi_gon, age, age_meth,  
-            SP_code, Chr_code_ger, bot_name, tpS_ID, LH_NH, H_SP_group, BWI_SP_group, Bio_SP_group, N_SP_group, N_bg_SP_group, N_f_SP_group_MoMoK,
-             DBH_class,  Kraft, C_layer, H_dm, H_m, H_method, C_h_dm, D_mm,   DBH_h_cm,  DBH_cm, BA_m2,
-            CCS_r_m, stand, stand_plot_A_ha, plot_A_ha)
+  # as there were some trees that had an estimated height which was lower then the DBH measuring height. this is not only implausible but also won´t work for TapeS 
+  # thus we correct these heights afterwards by estimating their height from the relation between the dg and hg and dg and the trees DBH (dreisatz, h_proportional function)
+  mutate(H_method = case_when(DBH_h_m > H_m & !is.na(H_g) ~ "ehk_sloboda_2",
+                              DBH_h_m > H_m & is.na(H_g) ~ "curtis_2", 
+                              TRUE ~ H_method),
+         H_m = case_when(DBH_h_m > H_m & !is.na(H_g) ~ ehk_sloboda(H_SP_group, DBH_cm*10, mean_DBH_mm, D_g*10, H_g*10),
+                         DBH_h_m > H_m & is.na(H_g) ~  h_curtis(H_SP_group, DBH_cm*10), 
+                         TRUE ~ H_m)) %>% 
+  # select columns that should enter the next step of data processing
+  # select(plot_ID, inv, inv_year, stand, tree_ID,  tree_inventory_status,  multi_stem, dist_cm,  azi_gon, age, age_meth,  
+  #        SP_code, Chr_code_ger, bot_name, tpS_ID, LH_NH, H_SP_group, BWI_SP_group, Bio_SP_group, N_SP_group, N_bg_SP_group, N_f_SP_group_MoMoK,
+  #        DBH_class,  Kraft, C_layer, H_dm, H_m, H_method, C_h_dm, D_mm,   DBH_h_cm,  DBH_cm, BA_m2,
+  #        CCS_r_m, stand, stand_plot_A_ha, plot_A_ha) %>% 
+  filter(!is.na(H_m))
 
 
  
 
-# 2.3.2. height calculation BZE -------------------------------------------------
+# 2.3.2. height calculation BZE --------------------------------------------------------------------------------------------------------------------------------------
 BZE3_trees_update_3 <-  trees_total %>% 
   filter(inv=="BZE3")%>% 
 ## 2.3.1. joining coefficients and Hg-Dg-data in 
@@ -268,15 +266,20 @@ BZE3_trees_update_3 <-  trees_total %>%
                          # and hm is na and the Slobody function cannot eb applied because there is no h_g calculatable use the curtis function
                          is.na(H_m) & is.na(R2_comb) & is.na(H_g)| is.na(H_m) & R2_comb < 0.70 & is.na(H_g) ~ h_curtis(H_SP_group, DBH_cm*10), 
                          TRUE ~ H_m))) %>% 
-    # as there were some trees that had an estimated height which was lower then the DBH measuring height. this is not only implausible but also won´t work for TapeS 
+  # as there were some trees that had an estimated height which was lower then the DBH measuring height. this is not only implausible but also won´t work for TapeS 
   # thus we correct these heights afterwards by estimating their height from the relation between the dg and hg and dg and the trees DBH (dreisatz, h_proportional function)
-  mutate(H_m = ifelse(DBH_h_m > H_m, h_proportional(d_g, H_g, DBH_cm), H_m))   %>% 
-     # select columns that should enter the next step of data processing
-     select(plot_ID, inv, inv_year, stand, tree_ID,  tree_inventory_status,  multi_stem, dist_cm,  azi_gon, age, age_meth,  
-            SP_code, Chr_code_ger, bot_name, tpS_ID, LH_NH, H_SP_group, BWI_SP_group, Bio_SP_group, N_SP_group, N_bg_SP_group, N_f_SP_group_MoMoK,
-            DBH_class,  Kraft, C_layer, H_dm, H_m, H_method, C_h_dm, D_mm,   DBH_h_cm,  DBH_cm, BA_m2,
-            CCS_r_m, stand, stand_plot_A_ha, plot_A_ha)
-
+  mutate(H_method = case_when(DBH_h_m > H_m & !is.na(H_g) ~ "ehk_sloboda_2",
+                              DBH_h_m > H_m & is.na(H_g) ~ "curtis_2", 
+                              TRUE ~ H_method),
+         H_m = case_when(DBH_h_m > H_m & !is.na(H_g) ~ ehk_sloboda(H_SP_group, DBH_cm*10, mean_DBH_mm, D_g*10, H_g*10),
+                         DBH_h_m > H_m & is.na(H_g) ~  h_curtis(H_SP_group, DBH_cm*10), 
+                         TRUE ~ H_m)) %>% 
+  # select columns that should enter the next step of data processing
+  # select(plot_ID, inv, inv_year, stand, tree_ID,  tree_inventory_status,  multi_stem, dist_cm,  azi_gon, age, age_meth,  
+  #        SP_code, Chr_code_ger, bot_name, tpS_ID, LH_NH, H_SP_group, BWI_SP_group, Bio_SP_group, N_SP_group, N_bg_SP_group, N_f_SP_group_MoMoK,
+  #        DBH_class,  Kraft, C_layer, H_dm, H_m, H_method, C_h_dm, D_mm,   DBH_h_cm,  DBH_cm, BA_m2,
+  #        CCS_r_m, stand, stand_plot_A_ha, plot_A_ha) %>% 
+  filter(!is.na(H_m))
 
 # 1.1.2.6. remove problematik trees ---------------------------------------
  # HBI
@@ -285,27 +288,27 @@ BZE3_trees_update_3 <-  trees_total %>%
                                            filter(DBH_h_cm/100 >= H_m | H_m >40 & H_method != "sampled") %>% 
                                            mutate(rem_reason = "LT excluded during height estimation"))
  # BZE3
+if(exists('BZE3_trees')){
  BZE3_trees_removed <- plyr::rbind.fill(BZE3_trees_removed, 
                                           BZE3_trees_update_3 %>%
                                             filter(DBH_h_cm/100 >= H_m | H_m >40 & H_method != "sampled") %>% 
                                             mutate(rem_reason = "LT excluded during height estimation")) 
+ }
  
 
-# ---- 1.1.2.6. exporting dataset --------------------------
+# ---- 2. exporting dataset --------------------------
  # height nls coefficients
-write.csv(coeff_H_comb, paste0(out.path, paste("coef_H", unique(HBI_trees_update_3$inv)[1], unique(BZE3_trees_update_3$inv)[1], sep = "_"), ".csv"), row.names = FALSE, fileEncoding = "UTF-8")
+write.csv(coeff_H_comb, paste0(out.path, "coef_H_", unique(HBI_trees_update_3$inv)[1], "_", unique(HBI_trees_update_3$inv)[2] , ".csv"), row.names = FALSE, fileEncoding = "UTF-8") ##change : we had to adjust this so it saves the coefficients as "hbi_momok"
                                
 # HBI dataset including estimated heights
 write.csv(HBI_trees_update_3, paste0(out.path, paste(unique(HBI_trees_update_3$inv)[1], "LT_update_3", sep = "_"), ".csv"), row.names = FALSE, fileEncoding = "UTF-8")
 write.csv(HBI_trees_removed, paste0(out.path, paste(unique(HBI_trees_update_3$inv)[1], "LT_removed", sep = "_"), ".csv"), row.names = FALSE, fileEncoding = "UTF-8")
 
 # BZE3 dataset including estimated heights
+if(exists('BZE3_trees')){
 write.csv(BZE3_trees_update_3, paste0(out.path, paste(unique(BZE3_trees_update_3$inv)[1], "LT_update_3", sep = "_"), ".csv"), row.names = FALSE, fileEncoding = "UTF-8")
 write.csv(BZE3_trees_removed, paste0(out.path, paste(unique(BZE3_trees_update_3$inv)[1], "LT_removed", sep = "_"), ".csv"), row.names = FALSE, fileEncoding = "UTF-8")
-
-
-
-#write.csv(HBI_trees_removed_3,  paste0(out.path, paste(unique(BZE3_trees_update_3$inv)[1], "weird_sloboda_heights", sep = "_"), ".csv"))
+}
